backend/requirements.txt
asgiref==3.4.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.5
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.0.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.1.1
environ==1.0
flake8==3.8.4
gunicorn==20.0.4
httpie==2.3.0
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.1
packaging==20.3
Pillow==8.1.0
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python3-openid==3.2.0
pytils==0.3
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
zipp==3.1.0
